+++ digiroad2-api-oth/src/main/scala/fi/liikennevirasto/digiroad2/Digiroad2Api.scala
@@ -67,7 +67,7 @@ class Digiroad2Api(val roadLinkService: RoadLinkService,
                    val assetPropertyService: AssetPropertyService = Digiroad2Context.assetPropertyService,
                    val trafficLightService: TrafficLightService = Digiroad2Context.trafficLightService,
                    val trafficSignService: TrafficSignService = Digiroad2Context.trafficSignService,
-                   val assetService: AssetService = Digiroad2Context.assetService,
+                   val complimentaryRoadLinkExpiringProcess: ComplimentaryRoadLinkExpiringProcess = Digiroad2Context.complimentaryRoadLinkExpiringProcess,
                    val verificationService: VerificationService = Digiroad2Context.verificationService,
                    val municipalityService: MunicipalityService = Digiroad2Context.municipalityService,
                    val applicationFeedback: FeedbackApplicationService = Digiroad2Context.applicationFeedback,
@@ -846,14 +846,16 @@ class Digiroad2Api(val roadLinkService: RoadLinkService,
   }
 
   delete("/roadlinks/complementaryLinksToDelete") {
+    val user = userProvider.getCurrentUser()
+    if (!user.isOperator()) {
+      halt(Unauthorized("User not authorized"))
+    }
     val linkIdsToDelete: Set[String] = Option(request.getParameterValues("linkIds"))
       .map(_.flatMap(_.split(",")).toSet)
       .getOrElse(Set.empty)
 
-    val username = userProvider.getCurrentUser().username
-
     roadLinkService.deleteComplementaryRoadLinksAndPropertiesByLinkIds(linkIdsToDelete)
-    assetService.handleAssetsOnDeletedComplementaryRoadLinks(linkIdsToDelete, username)
+    complimentaryRoadLinkExpiringProcess.handleAssetsOnDeletedComplementaryRoadLinks(linkIdsToDelete, user.username)
   }
 
   get("/roadLinks/incomplete") {
